mod : enum entity 내부에 설정 처리
