--- gusto/src/main/java/com/umc/gusto/domain/user/enums/Age.java
@@ -1,5 +1,0 @@
-package com.umc.gusto.domain.user.enums;
-
-public enum Age {
-    TEEN, TWENTIES, THIRTIES, FOURTIES, FIFTIES, OLDER, NONE
-}
