--- gusto/src/main/java/com/umc/gusto/domain/user/enums/Gender.java
@@ -1,5 +1,0 @@
-package com.umc.gusto.domain.user.enums;
-
-public enum Gender {
-    FEMALE, MALE, NONE
-}
