--- gusto/src/main/java/com/umc/gusto/domain/user/enums/MemberStatus.java
@@ -1,5 +1,0 @@
-package com.umc.gusto.domain.user.enums;
-
-public enum MemberStatus {
-    ACTIVE, INACTIVE, PAUSE
-}
