--- gusto/src/main/java/com/umc/gusto/domain/user/enums/SocialType.java
@@ -1,5 +1,0 @@
-package com.umc.gusto.domain.user.enums;
-
-public enum SocialType {
-    KAKAO, NAVER, GOOGLE
-}
